btrfs-fuse: introduce the skeleton for btrfs raid attrs
The naming is from kernel btrfs_raid_attrs, but the content is
completely different, as in btrfs-fuse we only care 2 things:

- How many mirrors the profile can provide

- The function to read the content

Unlike kernel, we don't really need all the complext parameters, we can
handle them all in the read function.

This patch introduce the skeleton for later implementation for each
profile.

Signed-off-by: Qu Wenruo <wqu@suse.com>