btrfs-fuse: data: fix a condtion in btrfs_lookup_csum()
There is a wrong condition in btrfs_lookup_csum() where it always
returns PTR_ERR(-ENOENT) when there is a csum covering the range.

This happens when the csum item end is beyond current bytenr, in that
case, we should return a pointer inside the item, but due to a wrong
condition we always reutnr PTR_ERR(-ENOENT).

This makes us to bypass all csum verification.

Signed-off-by: Qu Wenruo <wqu@suse.com>