btrfs-fuse: tests: disable subvolume boundary check in fssum
Since btrfs-fuse can not use st_dev to tell if we're at subvolume
boundary, fssum can report extra files only in btrfs-fuse, but not in
kernel, as it skips the subvolume.

To make it work properly for btrfs-fuse selftests, remove the subvolume
boundary check.

This can lead to some loop due to snapshots, but that's what we can
solve in the test cases.

Signed-off-by: Qu Wenruo <wqu@suse.com>