btrfs-fuse: fix a false alert on uninitialized variable
[BUG]
Fedora 36 reports the following uninitialized variable:

../data.c: In function ‘btrfs_read_data’:
../data.c:221:16: warning: ‘ret’ may be used uninitialized [-Wmaybe-uninitialized]
  221 |         return ret;
      |                ^~~
../data.c:133:13: note: ‘ret’ was declared here
  133 |         int ret;
      |             ^~~
../data.c: In function ‘btrfs_read_file’:
../data.c:539:17: warning: ‘ret’ may be used uninitialized [-Wmaybe-uninitialized]
  539 |         if (ret < 0 && cur_off == 0)
      |             ~~~~^~~
../data.c:505:13: note: ‘ret’ was declared here
  505 |         int ret;
      |             ^~~
../data.c:540:24: warning: ‘ret’ may be used uninitialized [-Wmaybe-uninitialized]
  540 |                 return ret;
      |                        ^~~
../data.c:440:13: note: ‘ret’ was declared here
  440 |         int ret;
      |             ^~~

[CAUSE]
This is in fact a false alert.

@ret will not be initialized only when we go through the csum lookup,
and reached the for () loop to iterate all copies.

Then if @max_mirror is 0 or smaller, we will not initialize @ret and
cause the warning.

But @max_mirror() is assigned from btrfs_num_copies(), which either
return <0 for error (already handled), or fetching the max_mirror number
from btrfs_raid_array[], where all max_mirror numbers are >= 1.

So this is a false alert.

[FIX]
Nevertheless, initialize @ret to 0 to silent the compiler which is not
able to dig deep enough.

Signed-off-by: Qu Wenruo <wqu@suse.com>