btrfs-fuse: allow readdir to utilize the offset parameter
The existing implementation of readdir() is passing @offset as 0 for the
filler() function.

This means we have to read in all the dir contents, this is not going to
handle large directories well.

Normally for readdir() on large directory, readdir() is implemented that
it can resume from the last read entry, so filler() won't need to
allocate a huge memory to contain all the info.

Add such ability to properly pass the found directory index into
filler(), and since the @offset passed into readdir() is the last found
entry, we need to pass @offset + 1 for btrfs_iterate_dir_start().

With this change, btrfs_fuse_readdir() is able to resume from the last
found entry, and handle huge directory correctly.

Issue: #11
Signed-off-by: Qu Wenruo <wqu@suse.com>